refactor: remove tw usage and use mui padding
